% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_map.R
\name{bg_map}
\alias{bg_map}
\title{Background Map for phosphorus variable plotting}
\usage{
bg_map(lon, lat, data, bg)
}
\arguments{
\item{lon}{longitude column name}

\item{lat}{latitude column name}

\item{data}{dataframe with georeferenced data, 
containning at least latitude and longitude columns}

\item{bg}{sf object representing a shapefile for background plot.
e.g. mountains, rivers, ecosystems}
}
\value{
ggplot object with plotted map
}
\description{
Background Map for phosphorus variable plotting
}
\examples{
data(mountain)

bg_map("Long", "Lat", data = ISRIC_P$georef, bg = mountain) 

}
