% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickP.R
\name{brickP}
\alias{brickP}
\title{Read Soil Phosphorus Variables raster from ORNL 2013}
\usage{
brickP(nc_file)
}
\arguments{
\item{nc_file}{ORNL 2013 netCDF file of Soil Phosphorus Variables}
}
\value{
raster brick of of ORNL Soil Phosphorus variables
}
\description{
Read Soil Phosphorus Variables raster from ORNL 2013
}
\examples{
nc_in <- "inst/extdata/GLOBAL_PHOSPHORUS_DIST_MAP_1223/data/pforms_den.nc"

phospho_brick <- read_P_ORNL(nc_in)

}
