% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_elevation.R
\name{classify_elevation}
\alias{classify_elevation}
\title{Convert Elevation values to classes}
\usage{
classify_elevation(elevation, breaks = c(0, 1000, 2000, 9000),
  labels = c("Low", "Mid", "High"))
}
\arguments{
\item{elevation}{vector of numeric elevation values}

\item{breaks}{numeric vector of class limits}

\item{labels}{character vector of elevation class labels}
}
\value{
factor of elevation classes converting \code{NA}s to \code{"Missing"}
}
\description{
Convert Elevation values to classes
}
\examples{
accn.info$Elevation_class <- classify_elevation(accn.info$Elevation)

}
