% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_P_ISRIC.R
\name{extract_P_ISRIC}
\alias{extract_P_ISRIC}
\title{Extract Soil Phosphorus Retention Potential from georeference coordinates}
\usage{
extract_P_ISRIC(rc, georef, soilclass, lon = "lon", lat = "lat")
}
\arguments{
\item{rc}{ratified raster with P retention class attributes}

\item{georef}{dataframe with georeferenced data, 
containning at least latitude and longitude columns}

\item{soilclass}{dataframe with soil P retention class table}

\item{lon}{longitude column name}

\item{lat}{latitude column name}
}
\value{
list containning:
\describe{
  \item{breaks}{vector of breaks for ggplot legend}
  \item{pal}{named vector of colors for ggplot legend/map as in ISRIC 2011}
  \item{georef}{dataframe of soil P retention potential main class added 
                to georef input}
}
}
\description{
Extract soil phosphorus retention potential main class from georeference
coordinates
}
\examples{
ISRIC_P <- extract_P_ISRIC(rc, accn.info, P_soil_class,
                          lon = "Long", lat ="Lat")

}
