% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_P_scales.R
\name{get_peaks}
\alias{get_peaks}
\title{function for this:
which.max(density(!!!sym(scale_x))$y
Nasty!}
\usage{
get_peaks(df, scale_x = "LD2", scale_y = "combined")
}
\description{
function for this:
which.max(density(!!!sym(scale_x))$y
Nasty!
}
