% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_soil.R
\name{is_soil}
\alias{is_soil}
\title{Soil class check}
\usage{
is_soil(x, reserve = 5)
}
\arguments{
\item{x}{factor or character vector with soil phosphorus retention classes}

\item{minp}{Minimum quasi probability for entropy estimation}
}
\description{
Soil class check
}
\examples{

}
