% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ISRIC2011.R
\docType{data}
\name{ISRIC2011}
\alias{ISRIC2011}
\title{Global Distribution of Soil Phosphorus Retention Potential. 
International Soil Reference and Information Centre, 2011.}
\format{An object of class \code{"RasterLayer"}; see \code{\link[raster]{raster}}.}
\source{
\href{http://www.isric.org/sites/default/files/isric_report_2011_06.pdf}{ISRIC}
}
\usage{
data(ISRIC2011)
}
\description{
Generalised GIS map of 16 main phosphorus retention potential classes
at 5 arc-minute resolution.
}
\details{
Acidity (pH), soil mineralogy, and clay content were used to rate
the inferred capacity for P retention using four classes (Low, Moderate,
High, and Very High) for each FAO soil unit combining them into 16 different
possible classes per map unit.

Each map unit of 5 by 5 arc-minutes consists of up to eight different soil
units. The overall soil phosphorus retention potential was assessed for each 
mapping unit, taking into account the P ratings and relative proportion of 
each component FAO soil unit. A hierarchical clustering of the combined soil 
unit ratings resulted in 16 main P retention classes (MainCLASS) subidivided 
into 166 2nd level classes (FullClass), and 280 3rd level classes. Only the
most  extensive and most limiting  P retention class (MainCLASS)  has been
shown in the generalised GIS map.  Classes like 'VH3'', Hi3', 'Mo3' and Lo3'
are typically comprised of a combination of several 2nd level P retention
classes, albeit in widely varying proportions.

Although uncertainties remain high, the analysis provides an
approximation of world soil phosphorus retention potential.

Values assigned assuming that the following is an ascending order
of soil phosphorus retention classes:

\tabular{rrl}{
ascending \tab main \tab description\cr
0 \tab WR1 \tab Oceans/Inland Waters\cr
1 \tab GL1 \tab Glaciers\cr
2 \tab RK1 \tab >75\% Rock Outcrops\cr
3 \tab RK2 \tab 50-75\% Rock Outcrops\cr
4 \tab Lo1 \tab >75\% Low\cr
5 \tab Lo2 \tab 50-75\% Low\cr
6 \tab Lo3 \tab 25-50\% Low\cr
7 \tab Mo3 \tab 25-50\% Moderate\cr
8 \tab Mo2 \tab 50-75\% Moderate\cr
9 \tab Mo1 \tab >75\% Moderate\cr
10 \tab Hi3 \tab 25-50\% High\cr
11 \tab Hi2 \tab 50-75\% High\cr
12 \tab Hi1 \tab >75\% High\cr
13 \tab VH3 \tab 25-50\% Very High\cr
14 \tab VH2 \tab 50-75\% Very High\cr
15 \tab VH1 \tab >75\% Very High
}
}
\examples{
data(ISRIC2011)
plot(ISRIC2011)
}
\references{
Batjes NH 2011. Global distribution of soil phosphorus retention. 
            Report 2011/06, Plant Research International (PRI), 
            Wageningen UR, and ISRIC – World Soil Information,
            Wageningen, 42 p. with dataset 
                 
(\href{ http://www.isric.org/sites/default/files/ISRIC_Report_2011_06.pdf}{ISRIC})
}
\keyword{datasets}
