% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_from.R
\name{layers_from}
\alias{layers_from}
\title{Reclassifies raster from RAT columns}
\usage{
layers_from(ratified, cols = NULL)
}
\arguments{
\item{ratified}{ratified raster object}

\item{cols}{vector of columns.}
}
\value{
list of reclassified raster objects
}
\description{
Reclassifies raster from RAT columns
}
\examples{

}
