% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mountain.R
\docType{data}
\name{mountain}
\alias{mountain}
\title{A global inventory of mountains for bio-geographical applications}
\format{An object of class \code{"sf"}; see \code{\link[sf]{sf}}.}
\source{
\href{https://ilias.unibe.ch/goto_ilias3_unibe_cat_1000515.html}{GMBA mountain inventory_V1.1}
}
\usage{
data(mountain)
}
\description{
Derived from shape file
}
\examples{
data(mountain)
}
\references{
Körner C, Jetz W, Paulsen J, Payne D, Rudmann-Maurer K,
            Spehn EM (2017)
            A global inventory of mountains for bio-geographical applications
            Alpine Botany 127(1): 1-15, DOI: 10.1007/s00035-016-0182-6 

(\href{https://link.springer.com/article/10.1007/s00035-016-0182-6}{Springer})
}
\keyword{datasets}
