% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_plot.R
\name{nb_plot}
\alias{nb_plot}
\title{Raster plot adjusted to rendering in R notebooks}
\usage{
nb_plot(r, axis.args = list(cex.axis = 2), ...)
}
\arguments{
\item{r}{raster object to be plotted.}
}
\value{
plot
}
\description{
Raster plot adjusted to rendering in R notebooks
}
\examples{
data(ISRIC2011)
nb_plot(ISRIC2011$main)
}
