% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat.R
\name{rat}
\alias{rat}
\title{get raster attribute table}
\usage{
rat(x)
}
\arguments{
\item{x}{ratified raster}
}
\value{
matrix with named columns as soil units and values as percentage,
        excluding miscealaneous soil units RK WR GL  and corresponding
        map units
}
\description{
get raster attribute table
}
\examples{
get_soil_composition(mapunit, FAO74)

}
