% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ISRIC_AT.R
\name{read_ISRIC_AT}
\alias{read_ISRIC_AT}
\title{Read Soil Phosphorus Retention Raster Attribute Table from ISRIC 2011 mdb}
\usage{
read_ISRIC_AT(mdb_file)
}
\arguments{
\item{mbd_file}{ISRIC 2011 geotiff}
}
\value{
list of dataframes:
#' \describe{
  \item{\code{FAO74}}{\code{KeyFAO74}, Soil Map of the world legend}
  \item{\code{mapunit}}{\code{P_RetMap_FINAL}, Map unit attributes including map unit
                 phophorus retention full class, main class,
                 soil unit composition, and dominant soil unit
                 phophorus retention class}
  \item{\code{soilunit}}{\code{P_newsuid_avg_PRET} soil unit phosphorus retention class,
                  physicochemical properties, Qs single value}
  }
}
\description{
Read Soil Phosphorus Retention Raster Attribute Table from ISRIC 2011 mdb
}
\examples{
extdata_dir <- system.file("extdata", package = "soilP", mustWork = TRUE)
mdb_file <- file.path(extdata_dir,
"Global_distribution_of_soil_phosphorus_retention_potential",
"ISRIC_Phosphorus_Retention_Potential.mdb")
RATs <- read_ISRIC_RAT(mdb_file)

FAO74    <- RATs$FAO74
mapunit  <- RATs$mapunit
soilunit <- RATs$soilunit
}
