% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_P_ISRIC.R
\name{read_P_ISRIC}
\alias{read_P_ISRIC}
\title{Read Soil Phosphorus Retention Main Class Raster from ISRIC 2011 geotiff}
\usage{
read_P_ISRIC(tif, soilclass = NULL, is = NULL, becomes = NULL,
  filename = NULL)
}
\arguments{
\item{tif}{ISRIC 2011 geotiff}

\item{soilclass}{dataframe with soil P retention class table}

\item{is}{8 bit value integer vector corresponding to soil P retention
main classes}

\item{becomes}{8 bit value corresponding to new order for soil P retention 
main classes}

\item{filename}{output geotiff file name with P retention classes stored 
in new order}
}
\value{
ratified raster object with P retention classes as 8 bit values
}
\description{
Read Soil Phosphorus Retention Main Class Raster from ISRIC 2011 geotiff
}
\examples{
tif_in <- "inst/extdata/tif/P_retention_potential_main_grey.tif"
tif_out <- "inst/extdata/tif/P_retention_potential_main_grey_ascending.tif"
rc <-  read_P_ISRIC(tif       = tif_in,
                    soilclass = soilclass,
                    is        = "arcgis",
                    becomes   = "ascending",
                    filename  = tif_out)

}
