% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_entropy.R
\name{row_entropy}
\alias{row_entropy}
\title{estimated entropy}
\usage{
row_entropy(m, minp = 1e-06)
}
\arguments{
\item{m}{frequency matrix}

\item{minp}{Minimum quasi probability for entropy estimation}
}
\description{
estimated entropy
}
\examples{

}
