% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale256.R
\name{scale256}
\alias{scale256}
\title{Rescale to 8 bit integer, with reserved lower integers}
\usage{
scale256(x, reserve = 5)
}
\arguments{
\item{x}{factor or character vector with soil phosphorus retention classes}

\item{reserve}{first integers not corresponding to original scale}
}
\description{
Rescale to 8 bit integer, with reserved lower integers
}
\examples{

}
