% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_composition.R
\name{soil_composition}
\alias{soil_composition}
\title{Get Extended Map Unit Soil Composition Matrix from ISRIC RAT}
\usage{
soil_composition(mapunit, FAO74)
}
\arguments{
\item{mapunit}{map unit raster attribute table}

\item{FAO74}{legend key from read_ISRIC_RAT}
}
\value{
matrix with named columns as soil units and values as percentage,
        excluding miscealaneous soil units RK WR GL  and corresponding
        map units
}
\description{
Get Extended Map Unit Soil Composition Matrix from ISRIC RAT
}
\examples{
get_soil_composition(mapunit, FAO74)

}
