% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_in.R
\name{swap_in}
\alias{swap_in}
\title{Reclassifies raster with dataframe column}
\usage{
swap_in(x, rat = raster::levels(x)[[1]], from = "ID", to = NULL)
}
\arguments{
\item{x}{raster object}

\item{rat}{dataframe with Raster Atribute Table,
e.g. P retention potential main classes}

\item{is}{column name of input raster value.}

\item{becomes}{column name of output raster value}
}
\value{
ratified raster object with P retention classes as 8 bit values
}
\description{
Reclassifies raster with dataframe column
}
\examples{
tif_in <- "inst/extdata/tif/P_retention_potential_main_grey.tif"
tif_out <- "inst/extdata/tif/P_retention_potential_main_grey_ascending.tif"
rc <-  read_P_ISRIC(file      = tif_in,
                    rat       = soilclass,
                    is        = "arcgis",
                    becomes   = "ascending")

}
