% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ISRIC2011.R
\docType{data}
\name{ISRIC2011}
\alias{ISRIC2011}
\title{Global Distribution of Soil Phosphorus Retention Potential. 
International Soil Reference and Information Centre, 2011.}
\format{A list of \code{"RasterLayer"} objects;
        see \code{\link[raster]{raster}}.}
\source{
\href{http://www.isric.org/sites/default/files/isric_report_2011_06.pdf}{ISRIC}
}
\usage{
data(ISRIC2011)
}
\description{
List of raster objects representing various Phosphorus Retention Potential
associated variables.
}
\examples{
data(ISRIC2011)
plot(ISRIC2011)
}
\references{
Batjes NH 2011. Global distribution of soil phosphorus retention. 
            Report 2011/06, Plant Research International (PRI), 
            Wageningen UR, and ISRIC – World Soil Information,
            Wageningen, 42 p. with dataset 
                 
(\href{ http://www.isric.org/sites/default/files/ISRIC_Report_2011_06.pdf}{ISRIC})
}
\keyword{DSMW}
\keyword{FAO74,}
\keyword{datasets,}
